U113-006 move copying of ALI files to gnatprove
Previously, gprbuild was copying ALI files from the object dir to
the library dir in the case of a library project. We don't want this
copy to happen in phase 2, only phase 1, but there is no easy way to
tell gprbuild in which phase we are. We therefore do this copy in
gnatprove instead, which is already copying ALI files for other
reasons.

* gnatprove.adb
(Copy_ALI_Files): copy files from to the lib dir as well

Change-Id: Ic684eb649df700539b6f31f2367f4e99ee0b6293