U112-018 change guard into assertion for Is_Unchecked_Union wrapper
The frontend Is_Unchecked_Union routine can be only called on
types (otherwise it will crash in assertion). So is its backend wrapper now.

This is because the concept of unchecked union is only applicable to types.

* src/why/spark_atree-entities.ads
* src/why/spark_atree-entities.adb
(Is_Unchecked_Union): Replace a defensive guard in body with a precondition
in spec.

Change-Id: I7f1e78c58393211d912971c9c87d71cb22515564