TB19-012 non-volatile only at top level
We implement the rule that taking the address of a non-volatile
variable is only allowed at the top level of an address clause.

We implement this rule as a legality check (check during marking) as
opposed to a static proof check, as this is a refinement of a
previously existing legality check, which allowed taking the address
of *any* object only at the top level of an address clause. The
check that an object whose address clause is not just the address of
another object must be volatile, has previously been implemented as
a static proof check, because implementing it as a legality check
was deemed too restrictive.

* spark_definition.adb
(Mark_Attribute_Reference): if no other violation was found, check
  if we take the address of a non-volatile object while not being at
  the top level of an address clause.

* tests/PA24-008__address: new test
* tests/U113-026__complex_address/arrays.ads: mark some variables as
    volatile so that the test still works as intended

Change-Id: I8b742adba584a86ba82fb1a72248512bb9d9964e