U205-005 assume dynamic invariant of borrowers at end after the update
We should only assume the dynamic invariant of the value of the borrowed
object at the end of the borrow after all checks are performed for the
assignment. Indeed, in this case, we cannot create an inconsistency, as
the current value of the borrower is known to realize the assumption.

* src/why/gnat2why-expr.adb
(Assignment_Of_Obj_Decl): Update the values at end after the assignment.
(New_Update_For_Borrow_At_End): Do not assume the dynamic invariant of
 the borrower at end, as the update for reborrow necessarily needs to
 happen before the assignment in reborrows, since it uses the value of
 the borrower before the reborrow to get the values of frozen components.
 Also do not assume the dynamic invariant of the borrowed object for
 reborrows, as it does not make sense, since it is a reborrow, not a
 borrow.
(Transform_Assignment_Statement): Assume the dynamic invariant of the
 borrower at end after a reborrow since it is no longer done during the
 update of the values at end.

* tests/SC03-021__binary_search_trees
Update sessions.

* tests/U205-005__not_null_borrow
File testcase.

Change-Id: I55ce61622dc2fd4f0d6158660688b508682ba90c