U208-007 generate frame condition for local borrowers
When we do a reborrow inside a loop, we modify the predicted value
of the borrower. We assume in the frame condition of the loop the
dynamic invariant of this predicted value and that its is_null field
is the same as the is_null field of the borrower.

* src/why/gnat2why-expr-loops-inv.adb
(Generate_Frame_Condition): If a local borrower is reborrowed inside
 the loop, we assume the dynamic invariant of its predicted value and
 that the is_null fields correspond.

* tests/SC03-021__binary_search_trees
Update session test.

* tests/U208-007__frame_for_borrowers
New test.

Change-Id: Ic05e4cf348700d80d899f323796bbaabf25b8a01