U204-021 fix crash on move for types with relaxed initialization
Use the init wrapper type for the __moved_relation predicate if the
access type might contain parts with relaxed initialization. Also try
harder to preserve the Ada node for arrays in split form, so as to
be able to retrieve the bounds.

* src/why/gnat2why-types.adb
(Create_Move_Function): If the designated type might contain parts with
 relaxed initialization, use the init wrapper type for the parameter of
 the __moved_relation predicate.

* src/why/gnat2why-expr.adb
(Insert_Move_Of_Deep_Parts): When introducing temporaries for parts that
 should be moved, preserve the Ada node if the expression is an array in
 split form.
(Move_Expression): Insert a conversion to the expected type before
 calling the move function. It is necessary as the Relaxed_Init flags of
 the expected and actual types might not match.
(Transform_Assignment_Statement): If the initial value is an array in split
 form, add an empty label node to provide the node of the lefthand side as
 the Ada node from which to retrieve the bounds.

* src/why/why-gen-binders.adb
(Get_Ada_Type_From_Item): The formal of a __move function does not have an
 Ada node. In this case, retrieve the type from the why type when possible.

* tests/S117-008__pointer/test.out
Minor change in counterexample.

* tests/U204-021__relaxed_access
New test.

Change-Id: I66b693d8a61d2a3a5d6f549006614e4ac44478df