U209-014 reject overlays and unchecked conversion on deep types
* src/spark/spark_definition.adb
(Mark_Address): Reject address clause on objects of a deep type.
(Mark_Attribute_Reference): Reject the 'Access attribute if its prefix
 is an object of a deep type. Also reorganize checks to avoid the need
 for the Found_Violation flag and reject address clauses that we cannot
 handle safely (on expressions which are not parts of volatile objects
 or complete non-volatile objects).
(Mark_Subprogram_Entity): Reject instances of unchecked conversion if
 either the source or the target type are deep.
(Mark_Entity): Mark the address clause after the entity so that we can
 assume the entity type has been marked while marking the address 
 clause.

* tests/QA11-012__addr/test.out
The test is rejected until we can handle 'Address on a part of a
non-volatile object.

* tests/U113-026__complex_address/test.out
New error messages.

* tests/RA26-007__overlay/arrays.ads
Make the array object volatile to avoid a violation.

* tests/U209-014__access_address
New test.

Change-Id: Ibfa3d60c72f5b8105664d6447703a0fa89d035ad