U126-015 Move logic for dealing with inlined subprograms inside gnat2why
SPARK plugin for GNAT Studio was previously calling semantic queries to
check for the presence of a contract on the spec/body of the current
subprogram for which analysis is requested, as well as the location of the
subprogram spec (in spec file or body file), as this influences the decision
to inline the subprogram in gnat2why.

This was partial, and more importantly not robust as semantic queries are
not meant to work from the LAL env in SPARK plugin.

Now move that logic to gnat2why, so that the plugin can simply add
--limit-subp in all cases.

[changelog]

* src/spark/gnat2why_opts-reading.ads (Limit_Subp): Update comment.
* src/spark/spark_definition.adb (Mark_Subprogram_Entity): Add logic to deal
 with the case where the subprogram is inlined for proof.
* src/spark/spark_util-subprograms.adb (Get_Body_Entity): Fix for entry
 families.
* src/spark/spark_util.adb (Int_Image): New query function.
* src/spark/spark_util.ads (Int_Image): New query function.
* testsuite/gnatprove/tests/N425-007__local_subp_no_contract/test.out
Welcome side-effect of the change.

Change-Id: I352548584355f0376c4483394cfccf17dc870b20