U210-056 add test with missing checks for use of uninitialized value
Missing checks when an uninitialized data is accessed both in the compound
statement itself (e.g. in the expression of the CASE statement) and in its
child statements (e.g. in statements of the CASE alternative).

The problem only occurs in ELSIF/CASE/WHILE; it doesn't occur in IF, which
is already special cased, nor in FOR, whose expression is captured in a
local constant just before the loop (because the loop statements can modify
object referenced in the loop expression).

* testsuite/gnatprove/tests/U210-056__flow_first_use/
Add; missing checks.

Change-Id: I79123aba5501a661e22d910df9269082fae1ec5c