P511-045 fix locations on checks for reads of uninitialized objects
Previously the Traverse_Func frontend routine was traversing the AST by
roughly following the syntactic ordering of child nodes, even though this
behaviour was never promised and depended on manual layout of the child
nodes in the parent.

Now this routine traverses the AST in arbitrary order (most likely
alphabetically). This breaks the First_Variable_Use flow routine, which is
supposed to find the first use of object within an arbitrary statement.

This patch recovers the previous behaviour of First_Variable_Use (and in
fact fixes a bit this routine), by only traversing into conditions of ELSIF
and WHILE statements while ignoring their lists of statements (they will be
traversed by another call to First_Variable_Use, if necessary).

* src/flow/flow-analysis.adb
(First_Variable_Use): Only examine objects referenced in the ELSIF and WHILE
statements themselves, not in their lists of statements.

* testsuite/gnatprove/tests/N424-011__flow_error_placement/test.out
Check now appears in a WHILE loop condition, not in its body.

Change-Id: I689467996fe407f33e191bdc14a39aa82c06a885