TA21-008 Fix crash on permission mismatch in loop
When issuing an error related to permissions at the end of a loop not being
at least as much as the permissions at the start of the loop, parameters for
the call to Perm_Error_Loop_Exit were inverted. Now fixed.

[changelog]

* src/why/gnat2why-borrow_checker.adb
(Check_Loop_Statement.Check_Is_Less_Restrictive_Tree): Fix order of
 parameters.

* testsuite/gnatprove/tests/T129-043__move_pointer/test.opt
* testsuite/gnatprove/tests/T129-043__move_pointer/test.out
Test passes now.

Change-Id: Ia3075e8215437d64ecbe25fac351f1bd8b0718c2