(no-tn-check) disable CVC4 because of huge memory hunger
Disable CVC4 on a test where it takes almost 5 minutes and 5GB of memory.
Z3 gives the same results in seconds and with almost no memory consumption.

The history of this test is a sequence of SKIP/XFAIL/timeouts, so we don't
appear to loose anything.

* testsuite/gnatprove/tests/N625-016__proof_massive_aggregate/
Disable CVC4 for proof and for counterexamples.

Change-Id: I5d7ce1376f18690d8d9f5fa98e8c2298f63f496e