U127-040 find Enclosing_Declaration for plugin
The GNATstudio plugin was crashing on subunits because semantic queries
(such as p_semantic_parent) are not currently supported by its more
limited libadalang environment.

We now perform the semantic query in GNATprove, so that the plugin can
work for subunits.

* src/flow/flow.adb
(GG_To_JSON): find the line, column of the enclosing declaration

Change-Id: Idacd06c5e968873a24a8d8b6291051e8bd626c00