U225-014 do not restore borrow context after nested package body
In the borrow checker, we should not restore the context after the
verification of nested package bodies. Indeed, deep variables might
have been declared in these bodies, and they need to be added to
the envoironement of the enclosing unit.

* src/why/gnat2why-borrow_checker.adb
(Check_Package_Body): Do not restore context at the end.

* tests/U225-014__borrow_nested_package
Very minimized reproducer.

Change-Id: Iac37c7a47e67f3eee5a60f82b92d73b8e22a4208