U202-028 improve detection of volatile expressions
Routine Is_OK_Volatile_Context, which is used in marking to detect
allocators and calls to volatile functions in restricted contexts,
is now more reliable. Its new parameter Check_Actuals is set to
True in marking, because they happen after all subprogram calls
have been resolved and the types of formal parameters corresponding
to actual parameters are known.

* src/spark/spark_definition.adb
(Mark): Adapt call to Is_OK_Volatile_Context for allocators.
(Mark_Call): Likewise for calls to volatile functions.

* testsuite/gnatprove/tests/OB12-011__polluting_abstract_state/test.out
Call to volatile function in default actual parameter is now detected.

* testsuite/gnatprove/tests/OB17-003__volatile_actual_call/test.out
Change in cause of an expression function being not in SPARK.

* testsuite/gnatprove/tests/OB17-003__volatile_actual_object/test.out
Now the message emitted by the frontend for volatile objects as
actuals is less precise (but the code is much simpler).

* testsuite/gnatprove/tests/S109-007__illegal_pointers/test.out
Allocator as an actual parameter is now rightly rejected.

Change-Id: I40d3e067ffe5e811eec71a23b32c94ac0dbbec0c
Depends-On: Idbc55a8d5ac14a94a5407cdd81bd2b3b5748fb8a