U301-042 fix crash in initialization-via-loop heuristic
In the flow heuristic for initialization-via-loop we attempted to handle
various CFG vertices where writes to array components might be attached.
These vertices actually come either from assignment statements or from
writable actuals of subprogram calls, i.e. actuals corresponding to
formal parameters of modes OUT or IN OUT (and of mode IN, if the formal
parameter is has an access-to-variable type).

The actuals of with array components were wrongly assumed to be either
indexed components or slices. However, they might also be selected
components, when writing a record component of an array component, e.g.:

  Set (X (J).C);

In general, they might be any "names" (in the sense of Ada RM 4.1), e.g.
type conversions like:

  Set (Integer (X (J)));

or calls to instances of Ada.Unchecked_Conversion, which were causing
similar crashes.

* src/flow/flow-control_flow_graph.adb
(Check_Defined): Replace a too strict case statement with a more relaxed
condition; in Get_Array_Index we only examine N_Indexed_Component nodes
anyway.

* testsuite/gnatprove/tests/U301-042__flow_loop_init/
Enable; crash is gone.

Change-Id: I4aa157464e4bc6323eadb7f0fb75cc7c27043d27