U218-042 add test with indirectly nested packages
This test checks how the Initializes contract is generated with
indirectly nested packages.

To generate the Initializes contract for P1 we should examine the
Initializes contract not just for P2 but also for P3 (though actually
the generated Initializes for P2 should mention P3).

* testsuite/gnatprove/tests/U218-042__flow_indirectly_nested/
Add; warning is spurious; it comes from empty generated Initializes for
the outermost package P1.

Change-Id: I00dde7399a2025dea12944b64dace4288bb26b98