U202-028 reject references to volatile objects in marking
Frontend rejected eferences to volatile objects from code with
SPARK_Mode => On. Now marking rejects such references from code with
SPARK_Mode => Auto.

* testsuite/gnatprove/tests/OC10-031__discriminant_as_default_arg/
A corner case with illegal reference to discriminant of a protected type
appeared in the default parameter of a protected subprogram. This
reference was safe (because the discriminant is immutable), but the
current SPARK RM doesn't allow it.

* testsuite/gnatprove/tests/U202-028__volatile_auto/
Primary test for this change; now rejected as expected.

Change-Id: Iad016daf7d8161c42adf98618d296688822de794