U315-004 add examples of loops in UG
* docs/ug/en/source/loop.rst
Add an example using Relaxed_Initialization and several using a
pointer-based datastructure.

* testsuite/gnatprove/run-tests
UG tests prefixed with `ug__long__` need more steps to prove.

* tests/*/loop_types.ads
Add types for pointer-based lists.

* tests/ug__init_access_list_zero
* tests/ug__init_access_list_index
Initialization loop on a pointer-based list.

* tests/ug__init_arr_zero_relaxed_init
Initialization loop on an array with Relaxed_Initialization.

* tests/ug__map_access_list_incr
Mapping loop on a pointer-based list.

* tests/ug__long__map_access_list_incr_copy
Mapping loop on a pointer-based list using a copy to express the post.

* tests/ug__search_access_list_zero
Search loop on a pointer-based list.

* tests/ug__validate_access_list_zero
Validation loop on a pointer-based list.

* tests/ug__long__update_access_list_zero
Update loop on a pointer-based list using a copy to express the post.

* tests/ug__search_access_list_max
Maximum loop on a pointer-based list.

Change-Id: I1244b0196011bece87261cc4c9472df089dffcd8