U315-027 re-observe and re-borrow are ok when the source is observed
We no longer reject an assignment directly inside an anonymous
access-to-object type when it is only observed. Note that this is
not enough to allow reborrows when the borrower is observes as the
source is still checked to be in the Unrestricted state. We could
also change that, but it is consistent with the RM, so it is left
at that for now.

* src/why/gnat2why-borrow_checker.adb
(Check_Statement):  Local observers and borrowers can always be
 assigned, unless they are themselves borrowed.

* tests/U315-027__find_max
New test.

Change-Id: Ic91fb72d832de16f47db9e2b160f4748938b49ce