U301-038 Add explanation for RTE in misplaced 'Old/'Loop_Entry
Because 'Old/'Loop_Entry are unconditionally evaluated on entry to the
subprogram or loop, it may be confusing to use them inside potentially
unevaluated expressions. This use is in fact forbidden in Ada and
allowed in GNAT with pragma Unevaluated_Use_Of_Old(Allow).

When an RTE check is unproved inside such an attribute reference,
issue a "possible explanation" part of the check message to clarify
to the user that the expression is unconditionally evaluated.

[changelog]

* src/flow/flow_error_messages.adb
(Get_Explanation): New function to create the final explanation.
(Error_Msg_Proof): Call the new function Get_Explanation.

* testsuite/gnatprove/tests/O512-022__vscomp2014_partition_rte/test.out
* testsuite/gnatprove/tests/SC04-005__LSP/test.out
Update outputs.

* testsuite/gnatprove/tests/U301-038__misplaced_old/eval.adb
* testsuite/gnatprove/tests/U301-038__misplaced_old/eval.ads
* testsuite/gnatprove/tests/U301-038__misplaced_old/test.out
Add test.

Change-Id: Iab2f08835502dfe5eb7a17fe428206aa13dcecd2