U301-009 support for access-to-constant types in lrm
To recall, we currently reject:
 (1) Assignments to parts of a data-structure designated by an
     access-to-constant type (even nested under another
     access-to-variable type), both directly and through OUT parameters
     of procedure calls.
 (2) Borrowing of a part of a data-structure designated by an
     access-to-constant type
 (3) Assignments which should be moves (they contain access-to-variable
     parts) if the source is a part of a data-structure designated by an
     access-to-constant type and the target is not a part of a constant
     or the constant has an access-to-variable type.
 (4) Converting from an anonymous access-to-constant type to a named one
     to avoid bad interaction with observers.
In addition, (5) to avoid memory leaks, allocators returning an
access-to-constant type, and conversions from an access-to-variable type
to an access-to-constant type are only allowed in the declaration of
library-level constants. Both are considered to move their expression.

(1), (2), and (3) should be implicit if access-to-constant types are
considered to provide a constant view of the object.
(4) is an explicit legality rule
(5) is changed to be a tool limitation, we might want to allow these
construct and instead emit a (necessarily failed) verification
condition for them.

* source/declarations-and-types.rst
Update RM wording for access types.

* src/spark/spark_definition.adb
(5) is now a tool limitation.

Change-Id: Ie5cfe7161d156b0d5d27d694955ac699dc2345c9