U319-015 address clauses, import, and default initialization
Reject declarations of objects with an address clause if they
are not either marked as imported or fully initialized at declaration
(implicitly or explicitly). If the object with the address clause is a
constant then it should be imported.

* src/spark/spark_definition.adb
(Mark_Address): For supported cases, reject constant objects not marked
 as imported and variables neither imported nor fully initialized at
 declaration. Also warn on unsupported cases if a constant is not
 imported that the effects of the initialization are ignored.

* tests/MC12-024__flow_crash
* tests/MC12-026__overlayed_objects
* tests/Q106-016__address_aliasing
* tests/Q303-009__volatile
* tests/Q809-001__address
* tests/QA11-012__addr
* tests/RA26-007__overlay
* tests/TB19-012__address_clauses
* tests/TB19-012__aliasing
* tests/U113-026__complex_address
* tests/U209-014__access_address
* tests/U301-040__annotate
Add Import aspects on objects with an address clause.

* tests/U319-015__overlay_init
New test.

Change-Id: Ice47c4ae01293d703932605aa2f9883725386542
Depends-On: Ic4bd58215c657274f8fabd0dbf9225770d5f822b