U316-010 add test with subprogram overlying an object
Add an perfectly reasonable contract to Set_Overlay_Alias, a test that
was violating this contract, and adjust marking to avoid this violation.

* src/spark/spark_definition.adb
(Mark_Address): Move call to Set_Overlay_Alias where we also check that
the overlay applies to an object.
* src/spark/spark_util.ads
(Set_Overlay_Alias, Overlay_Alias): Add contracts.
* testsuite/gnatprove/tests/U316-010__subp__addr/
Add test with subprogram overlaying an object.

Change-Id: I90677b03283bab10d9b3f16c07aab65cfe6c48ef