U323-001 use original component for comparison in component sets
Change the data structure used for components to use a comparison
which considers components with the same unique representative as
equivalent.

* src/why/gnat2why-tables.adb
(Lt_Components, Eq_Components): Comparison operators on entities. On
 components, go to a unique representative before the comparison.
(Component_Sets): Instance of Ordered_Sets using the comparison
 operators defined above.
(Record_Info): Change the type of Components to use Component_Sets.Set.
(Has_Duplicates): Return True if there is a component in a component set
 which has the same name as a given component.
(Search_Component_In_Info): Use Find to get a component with the same
 unique component as the input component.
(Search_Component_In_Types): Inline the call to Search_Component_In_Info
 to avoid copying the component set.

* src/counterexamples/gnat2why-counter_examples.adb
Remove unnecessary with clause.

* src/why/why-gen-records.adb
Update the types of the component set in renames.

* tests/R618-026__hidden_comp_in_counterex/test.out
Minor changes in counterexamples.

Change-Id: I81713e3b702a586664840c9ced7bea2ee92f21dd