U325-003 fix crash when checking for UC
A type may have a known Esize, but the bounds may still not be known
at compile time, as is the case for dynamic subtypes. Checking for
the bounds to be known at compile time before accessing their value.

* spark_uilt-types.adb
(Suitable_For_UC_Target): check for the bounds to be known at
  compile time before accessing them

Change-Id: Ifb857e77cec486b920f75c2ce588639334ceccfa