U316-010 fix aliasing of loop params
When considering aliases via address clauses, the code was not
considering loop parameters as being constant. We fix this by making
the function Is_Constant_In_SPARK aware of loop parameters (and in
parameters of functions while we are at it).

* spark_util.adb
(Is_Constant_In_SPARK): check for loop parameter and in parameters
  of functions

Change-Id: I16ed8eba912ff90608f88e7c810e5b75724a5b98