TB19-042 Create DDG only on live code
Previously, the DDG was created by iterating on all vertices of the
CFG. Because the ABEND edges are not detected at creation of the CFG,
flow was crashing when traversing exceptional paths in the DDG
creation. By traversing only live code, the crashes disappear and flow
output isn't changed.

* src/flow/flow-data_dependence_graph.adb
(Create): Process the vertices with a DFS on the Start_Vertex instead
of a loop on all vertices.

* testsuite/gnatprove/tests/TB19-042__flow_exceptional_dead/
Add test cases. Remove from XFAIL.

Change-Id: I8f4d679b85e0d7093a8d39f2b5a60fff791e1d91