TB19-013 allow 'Access on objects in the RM
* source/declarations-and-types.rst
Objects that are not managed can be borrowed and observed.

* source/names-and-expressions.rst
'Access is allowed on objects, but only where borrowers/observers are
allowed if it is of an anonymous access type, and only on constant parts
of objects otherwise (ie. if it is of a named access-to-constant type).

* spark/spark_definition.adb
(Mark_Attribute): Corner case, it is OK to borrow a constant part of a
 local observer.

* spark/spark_util.adb
(Traverse_Access_To_Constant): It is not an access-to-constant part if
 the dereference is anonymous, a part of an observer is not necessarily
 constant.

* tests/S905-002__pointer_observe/test.out
* tests/TB19-013__access_attr/test.out
Minor changes in formulation of violations.

Change-Id: I6b02e15c960f749c20cf5dae9c2b940fbc61225d