U402-029 add a sanity check for problem with generated globals
A common symptom of various problems in flow analysis is that we
complain about objects missing from the Global contract while the Global
contract is in fact generated. This is now reported in developer builds
as an assertion failure, to make sure that we detect such problems.

* src/flow/flow-analysis-sanity.adb
(Check_All_Variables_Known): Add a sanity assertion.

Change-Id: I6eed985b0ad035a4bb09ad104eb9ae11fcce8b3f