U413-031 Use correct value of -j for ALI generation phase
In the ALI generation phase, we are passing directly the project file
to GPRbuild, which may contain a value for -j which is not the one
specified on the command line. Always add the correct value of -j after
the project file argument.

In the flow and proof phase, the defect was not really visible as it
was countered by the use of semaphores limiting the calls to gnat2why.

The problem was only present if -j was set to 1 on the command line.

[changelog]

* src/gnatprove/gnatprove.adb (Call_Gprbuild): Force correct value of -j
which may come from the command line.

Change-Id: I0fd50179ac28a861e9c29e65b8c9a15fa989f634