U303-019 Use info message instead of warning for no SPARK_Mode on unit
When a unit is analyzed specifically using the -u switch, and this unit
does not have SPARK_Mode specified, a message is issued. This should be
an info message rather than a warning one, to avoid it being treated
as an error when --warnings=error is used.

[changelog]

* src/why/gnat2why-driver.adb (GNAT_To_Why): Turn that particular
message into an info message.

* testsuite/gnatprove/tests/MB13-068__no_analysis/test.out
* ../gnatprove/tests/R522-048__flow_crash_compiler_switch/test.out
* testsuite/gnatprove/tests/R629-018__warning_no_analysis/test.out
Update output.

Change-Id: I2e389a1490bd03f34ca523da05fbc456cc88a72a