U414-017 raise an error in the borrow checker on missing globals
The borrow checker relies on the Global contracts supplied by flow
analysis to set the environment when analysing a callable body. If this
contract is incorrect, an object might be missing from the environement.
If it is the case, abort borrow checking with an error message instead
of crashing. Alternatively, we could handle these missing globals as
Read_Only, but some messages might be missed if the subprogram with
the wrong Global contract is called on a moved or borrowed object
(we would still have the flow message though).

* src/why/gnat2why-borrow_checker.adb
(Current_Subp): Global variable set to the enclosing unit during the
 analysis of a callable body.
(Check_Callable_Body): Set Current_Subp.
(Get_Perm_Or_Tree_Ent, Process_Path): Raise an error if the object is
 not in the permission environment and it is not a constant without
 variable inputs.

* tests/U414-017__missing_deep_global
New test with incorrect global contracts.

Change-Id: Ib85bf1e54f8eb2679118cb16356d6981291927c8