U416-015 decouple proof labels from assumptions
Routine Subp_Location, which provides proof with labels, relied on the
assumption machinery. This was unnecessarily complicated and also
inconsistent with a similar Subp_Body_Location routine, which doesn't
rely on assumptions.

Finally, the assumption machinery was needlessly calling Source_Name on
itypes (where the call to Source_Name is not even needed for proof
labels).

* src/spark/spark_util-subprograms.adb
(Subp_Location): Rewrite without using the assumptions machinery.

Change-Id: If7cfbcaaf4b2fcb118ffac817ee5bc5da3354bcd