U402-026 Update testsuite outputs
* update-all-session-tests
Allow --rewrite switch passed on to update-session-test.

* update-session-test
Make script more robust.

* K406-013__formal_lists/test.out
* R509-011__higher_order_proof/test.out
* SC03-021__binary_search_trees/test.out
* T727-043__mergesort/test.out
* bitwalker/test.out
* prime_numbers/test.out
* railway_signaling/test.out
Changes in provers but not provability, following changes in session.

* pointer_based_maps/test.out
New checks on calls to allocating functions.

* N624-020__perm_rem/test.out
Loss of a few proofs. All but 2 are recovered with more steps and 3
provers, but such a setting is too long for the testsuite run.

* PC14-014__formal_containers/test.out
Changes in provers, and one less proved check.

* QB10-026__counterexample/test.out
Changes in counterexamples.

Change-Id: I87e4c652c10fd9be2f0e089533e615d7e4f6d402