U416-024 Update SPARK RM regarding abstract state and volatile functions
SPARK RM 7.1.3(8) is updated to reflect the fact that abstract states
which do have Async_Writers or Effective_Reads cannot have as
constituents objects which are effectively volatile for reading, hence
need not require that a function reading such an abstract state be
marked as a volatile function.

[changelog]
* docs/lrm/source/packages.rst
Update SPARK RM 7.1.3(8) to allow non-problematic case.
* testsuite/gnatprove/tests/U416-024__abstract_state/foo.adb
* testsuite/gnatprove/tests/U416-024__abstract_state/foo.ads
* testsuite/gnatprove/tests/U416-024__abstract_state/test.out
Add test.

Change-Id: Ib318234b7f18f53990f6f67583e811fda215133d