U511-034 assume dynamic invariant of result after refined post
Subprograms which have a Refined_Postcondition aspect are translated as
follows. The body is translated and stored in an abstract block with the
refined post as a postcondition. The actual postcondition of the
subprogram if there is one is proved from the refined postcondition.
Before checking the postcondition, it is necessary to assume again
typing information which might not be included in the refined post. It
was already done for effects of procedures but not for results of
functions.

* src/why/gnat2why-subprograms.adb
(Checking_Of_Refined_Post): For functions, assume the dynamic invariant
 of the result. For borrowing traversal functions, also assume the
 dynamic invariant of the prophecy variables for the borrower and
 borrowed paramater.

* src/spark/spark_definition.adb
(Mark_Call): Allow At_End_Borrow function calls in refined post.

* tests/U511-034__refined_post_at_end
New test.

Change-Id: I02a5c93d667dd3db38ae42c58ab90c7be4cd5383