TA21-015 Support statically known exceptions raised
Support constructs for which the frontend generates a raise-expression
or raise-statement in the AST, as the exception is statically known to
be raised. This allows supporting previously rejected codes, as well
as simplifying the handling of such cases in GNATprove mode inside
the frontend.

[changelog]
* src/flow/flow_error_messages.adb (Get_Severity): Issue high message
on such statically known exceptions raised.
* src/spark/spark_definition.adb (Mark): Handle the case of node kind
N_Raise_xxx_Error. Return after issuing an error on ill-formed AST due
to static exception.
* src/why/gnat2why-expr.adb (Transform_Expr): Handle the case of node
kind N_Raise_xxx_Error.
* testsuite/gnatprove/tests/OB11-016__simple/test.out
* testsuite/gnatprove/tests/M923-035__discr/test.out
* ../gnatprove/tests/P114-015__marking_infinite_recursion/test.out
Update output.
* testsuite/gnatprove/tests/TA21-015__static_raise_exception
Add test.
* testsuite/gnatprove/tests/S708-039__pointer_search
Update test by removing obsolete comment. If variable Current is not
marked as "not null" at declaration, the code is accepted but an
additional loop invariant is needed to prove that the access is safe.

Change-Id: I4cdae0c0ff0d0943edc0153274694f660e377e00
Depends-On: I65877532131f190a32f9d8368c0a902f2092d509