U531-002 remove trivial explanations
If the message "cannot prove X" just repeats an X from the source
with no splitting involved, we remove it. We achieve this by not
generating a pretty-printing label at the top-level of an assertion,
only at subsequent levels.

To do this, we go back to a three-valued flag for label generation,
indicating whether we are at the top-level of a term or not. Update
this flag accordingly in Transform_Expr.

Testsuite updates are just removal of such trivial explanations.

* gnat2why-util.adb
(GM_Kind): new 3-valued type to replace Boolean flag
* gnat2why-subprograms.adb
* gnat2why-types.adb
replace True/False for Boolean flag with the appropriate value from
the three-valued type
* gnat2why-expr.adb
(Transform_Expr): update flag after top level

Change-Id: Ie866a4d4585343a0f127619f966333519d2c08b9