minor enable analysis of Graphs package by CodePeer
Now that one problematic routine in the Graphs package is explicitly
skipped by CodePeer, we can enable analysis of packages where Graphs
are instantiated.

Also, the list units where Graphs package is instantiated was actually
inaccurate and the exclusion of the Graphs package was useless (because
CodePeer only analysis instances and not generic units).

no-tn-check

* gnat2why/gnat2why.gpr
(CodePeer): Remove all special casing for the Graphs package and its
instances.

Change-Id: I46c55ff438d8505c2a083b83315bbcbadefad8c7