U426-011 support 'Access of a general access type
This patch mostly updates the borrow checker to support 'Access inside
moves. We use two mechanisms for that. If the root is deep, then it is
in the permission map, we use it. Otherwise, if the enclosing unit is a
package, reject the move as occuring inside elaboration. Otherwise, the
enclosing unit is a subprogram or a concurrent type. If the root of the
move is local to the unit, handle it as a borrow (using a new
Shallow_Moves set). Otherwise, reject it stating that the root will be
moved at the end of the unit.

* src/spark/spark_definition.adb
(Mark_Attribute_Reference): Accept references to 'Access of a named
 access-to-variable type. Like for anonymous access types, check that
 they only occur in declarations, assignments, and return statements
 (though the last case will be rejected by either the frontend or the
 borrow checker).
(Mark_Type_Entity): E_Access_Attribute_Type can occur now, but only in
 rejected constructs ('Access inside an expression).

* src/spark/spark_util.adb
(Is_Path_Expression): Separate constructs that can occur anywhere inside
 a path from those that can occur only at top-level ('Access, 'Length,
 = null...).

* src/why/gnat2why-borrow_checker.adb
(Shallow_Moves): Set of all the local shallow paths which have been
 moved.
(Check_Not_Moved):  Check that a path rooted in a shallow root was not
 moved.
(Handle_Move_Of_Shallow): Handle a move of a path inside a shallow root.
(Set_Perm_Prefixes_Move_Access): Modifiy the permissions of a given node
 in the permission environment as well as all the prefixes of the path,
 to the permission No_Access until a dereference is found, and
 Write_Only afterward.
(Read_Indexes): Ignore calls to predicate functions if they are not in
 SPARK (as we do in marking and proof).
(Process_Path): Move handling of observe/borrow at the beginning of the
 process as it is done for both deep and shallow roots. Handle moves
 of shallow roots separately. Update the permission tree specifically
 for moved paths with a 'Access.

* src/why/gnat2why-expr.adb
(Transform_Assignment_Statement): Small bug fix - introduce a temporary
 if a tag check is needed to avoid generating incorrect Why3 code.

* tests/P831-045__pointers/test.out
* tests/TB19-013__access_attr/test.out
* tests/TU__access_attribute/test.out
Language violations caused by 'Access are replaced by violations of
ownership.

* tests/U426-011__create_gen_access
* tests/U426-011__general_access_attribute
New tests.

Change-Id: I0077a7c842c9f3b203d93bab1919fc8ca8dff207