(no-tn-check) allow toggling counterexamples per file
We fix parsing of --counterexamples switch, which was not allowed in
file-specific switches. We also improve the test driver to allow the
user to disable the global setting of the counterexample switch.

* configuration.adb
(Parse_Switches): move declaration of --counterexamples switch to
the right section
* test_support.py
(prove_all): allow setting counterexamples argument to None to not
  force any of --counterexamples=on or off

Change-Id: I9df81ba34d04c1091ff508e139dc744fc429ec87