U621-023 reject moves of objects that may have overlays
We reject a move of a part of X if either X is the overlay of something
else or if somewhere in the visible code, X has an overlay.

* src/why/gnat2why-borrow_checker.adb
(Handle_Move_Of_Shallow): Make sure that the root of the moved
 expression has no known aliases.
(Check_Assignment): Reformulate error message to make it clear it is a
 tool restriction.

* tests/U621-023__moved_overlays
File test.

* tests/U427-010__borrow_overlay/test.out
Modified error messages.

Change-Id: Iec082f05b3d8d6f74bcbab5cb556565fc7cbca59