U121-019 detect stability in loop conditions
We now detect when a loop condition is 'stable', i.e. where the loop
condition variables are not modified within the body of the loop. This
is an indication that the loop is non-terminating.

* src/flow/flow.adb
(Flow_Analyse_CUnit): Find stable conditions in subprograms and packages
 during flow analysis.

* src/flow/flow-analysis.adb
* src/flow/flow-analysis.ads
(Find_Stable_Conditions): Look through all the vertices in the control
 flow graph that pertain to a loop. If the variables used in the
 condition are not all defined within the loop body, then we issue a
 warning. Otherwise it is considered not stable i.e. terminating.

* src/flow/flow.ads
* src/flow/flow-control_flow_graph.adb
Remove Flow_Analysis_Graphs_Root.Loops field as it is no longer needed

* testsuite/../L305-015__actions/test.out
* testsuite/../M304-011__flow_contract_checking/test.out
* testsuite/../P715-019__flow_subprogram_termination/test.out
* testsuite/../Q313-001__renaming/test.out
* testsuite/../Q912-021__codepeer_by_example/test.out
* testsuite/../R423-012__dead_code/test.out
* testsuite/../SA16-025__inlining/test.out
* testsuite/../T618-021__explanations/test.out
* testsuite/../U121-019__flow_detect-loop-stability/test.out
* testsuite/../U210-056__flow_first_use/test.out
Update test output

* testsuite/../U121-019__flow_detect-loop-stability/q.adb
* testsuite/../U121-019__flow_detect-loop-stability/r.adb
Add new tests demonstrating detection within nested loops

Change-Id: I31868090e7553c72f193e2283dcd56ebfcef54e8